new changes

+++ Basic Python Question/PostiveNegativePgm.py
@@ -1,4 +1,4 @@
-num = float(input("Enter a number = "))
+num = int(input("Enter a number = "))
 
 if num > 0:
     print("number is positive")
